src/requirements.txt
python-telegram-bot==20.3
aiohttp==3.9.1
python-dotenv==1.0.0